refactored routes to declutter index file
